Complete A-MoreOOP
* Refactorization:
1) BotBrain -> Main Hub that pull in all Units together and react to human response
2) BotCommandResponderUnit -> Decipher command from user and produce the respective action
3) BotDynamicMemoryUnit -> Saving and Loading data.txt
4) BotStaticMemoryUnit -> Serve as a storage space for messages
5) BotTaskStatusGeneratorUnit -> Generate status of tasklist
6) BotTemporalUnit -> Incharge of converting String data to DateTime
