+++ src/main/java/Bot.java
@@ -1,8 +1,3 @@
-import java.time.Period;
-import java.util.Arrays;
-import java.util.Scanner;
-import java.util.stream.IntStream;
-
 public class Bot {
 
     /**
