+++ src/main/java/BotException.java
@@ -1,5 +1,4 @@
 import java.io.IOException;
-import java.time.format.DateTimeParseException;
 
 class InvalidCommandException extends Exception {
     public InvalidCommandException(String msg) {
