+++ src/main/java/Task.java
@@ -1,4 +1,3 @@
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 public abstract class Task {
